adding the meory detection and management

+++ kernel/src/includes/conversion.h
@@ -2,4 +2,4 @@
 #include "../includes/stdint.h"
 
 const char* toString(int num);
-const char* hex_to_String(uint32_t value);
+const char* hex_to_String(uint64_t value);
